+++ nboost/models/pt_models/distilbert_qa.py
@@ -63,7 +63,9 @@ class PtDistilBertQAModel(QAModel):
             start_logits = start_logits[0][len(truncated_query) + 2:-1]
             end_logits = end_logits[0][len(truncated_query) + 2:-1]
 
-        assert len(end_logits) == len(tok_to_orig_index)
+        assert len(end_logits) == len(tok_to_orig_index) or len(end_logits) == \
+               self.max_seq_len - len(truncated_query) - 3
+
         start_tok = int(np.argmax(start_logits))
         end_tok = int(np.argmax(end_logits[start_tok + 1:])) + start_tok