+++ nboost/proxy.py
@@ -376,8 +376,8 @@ class Proxy(SocketServer):
 
                 response['body']['nboost'] = {}
 
-                if self.qa:
-                    answer = self.qa_model.get_answer(query, cvalues[0])
+                if self.qa and len(cvalues) > 0:
+                    answer = self.qa_model.get_answer(query, cvalues[ranks.index(min(ranks))])
                     response['body']['nboost']['qa_model'] = answer
 
             self.client_send(request, response, client_socket)